TwpingJson: add timeout support
Add an optional timeout argument to TwpingJson.run() which allows
the caller to request that twping is terminated if it has not done
so before the timeout has elapsed.

This is implemented through the use of a threading.Timer instance.
If the timer expires it calls the terminate() function. _proc_lock
ensures there is no contention if the timer expires at the same time
as the process exits of its own accord.

VRVDR-43755